improve js refactoring with components done
