+++ scripts/components/renderUser/local.js
@@ -1,7 +1,7 @@
-import { updateUserData } from "./helper/updaters.js";
-import { userInput, body } from "./helper/querySelectors.js";
+import { updateUserData } from "./updaters.js";
+import { userInput, body } from "../helper/querySelectors.js";
 
-export const onloadLocalUser = () =>
+export const localRenderData = () =>
   (body.onload = () => {
     if (localStorage.getItem("template") && localStorage.getItem("user_id")) {
       userInput.value = localStorage.getItem("user_id");
